Rename formatters module
